# docs/swagger.yaml
basePath: /
definitions:
  transport.transort_error:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: AuthService
  version: "1.0"
paths:
  /auth/getTokens/:
    get:
      description: Generate tokens
      operationId: generate-tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.transort_error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.transort_error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.transort_error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/transport.transort_error'
      summary: Generate tokens
      tags:
      - Auth
  /auth/refresh/:
    post:
      description: Refresh
      operationId: refresh
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.transort_error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.transort_error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.transort_error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/transport.transort_error'
      summary: Refresh
      tags:
      - Auth
  /auth/register/:
    post:
      description: Register
      operationId: register
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.transort_error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.transort_error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.transort_error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/transport.transort_error'
      summary: Register
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"